requirements.txt
attrs==21.4.0
brotlipy==0.7.0
ConfigParser==6.0.1
cryptography
Cython==0.29.32
dl==0.1.0
docutils==0.18.1
docutils==0.18.1
Flask==3.0.2
fuzzywuzzy==0.18.0
HTMLParser==0.0.2
importlib_metadata==4.11.3
ipython==8.12.3
ipywidgets==7.6.5
Jinja2==3.1.3
keyring==23.4.0
lockfile==0.12.2
lxml==4.9.1
numpy==1.21.5
openai==1.13.3
ordereddict==1.1
pandas==1.4.4
Pillow==10.2.0
protobuf==3.19.6
pyOpenSSL==24.1.0
pyspark==3.5.0
railroad==0.5.0
rapidfuzz==3.6.1
scikit_learn==1.0.2
Sphinx==5.0.2
thread==1.0.0
toml==0.10.2
tornado==6.1
zipp==3.8.0
